requirements.txt
openai
# python-dotenv
langgraph
redis
scipy
pydantic
newsapi-python
python-docx